requirements.txt
# Function dependencies, for example:
# package>=version
requests
google-cloud-storage==1.43.0
SQLAlchemy==1.4.36
PyMySQL==1.0.2
